check valid RSA key
